+++ src/main/java/domain/CrewRepository.java
@@ -5,7 +5,7 @@ import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import util.FileParser;
+import util.parser.FileParser;
 
 public class CrewRepository {
