+++ src/main/java/util/parser/AttendanceRecord.java
@@ -1,4 +1,4 @@
-package util;
+package util.parser;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
